fix(rag_pipeline_2): 베이스라인 정밀도를 bfloat16에서 float16으로 변경함
- Tesla T4가 bfloat16을 하드웨어 가속하지 않아 float16으로 통일함
- vLLM 엔진 서버 실행 시 dtype 검증 에러로 확인됨

+++ src/rag_pipeline_2/generation.py
@@ -19,7 +19,7 @@ MAX_NEW_TOKENS = 512
 
 def load_model_and_tokenizer() -> tuple[PreTrainedModel, PreTrainedTokenizer]:
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
-    model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=torch.bfloat16, device_map="auto")
+    model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=torch.float16, device_map="auto")
     return model, tokenizer
